Malltiverse/feature: add payment screen

+++ Malltiverse/app/src/main/java/dev/borisochieng/timbushop/presentation/ui/components/CartItem.kt
@@ -11,7 +11,6 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
@@ -24,7 +23,6 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableDoubleStateOf
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
@@ -55,7 +53,8 @@ fun CartItem(
     onRemoveFromCart: () -> Unit,
     onQuantityChange: (DomainProduct, Int) -> Unit
 ) {
-    var itemTotalPrice by remember { mutableDoubleStateOf(product.price.toDouble() * product.quantity) }
+    var itemTotalPrice by remember { mutableDoubleStateOf(product.price * product.quantity) }
+
     Card(
         modifier = modifier
             .height(150.dp)
